Fixme (Remove 'True') - Add GraphQL starting point to the web server.
